feat(student-documents): adding table and input on page
